git: don't force external difft
Ok so this is a deep cut.
The problem is that when you have diff.external set, it overrides
*everything*.
So if you want to have, for example, a textconv diff filter… you just
can't do that.
Even passing --no-ext is questionable.
Instead, just require the use of `dt` / --ext-diff.
It's tragic :(