nvim: rework lsp module
There's still some improvements to be made, but this is cleaner.
We lose global caps defaults, but those made no sense to begin with.
We do also lose caps overrides, I'll add them back in if I need them.

In the process, we gain `every?` and `executable?`.

There's also a misc comp print vim.inspect to vim.print change.