sh: add /tmp helpers
I have really poor discipline when it comes to /tmp.
In particular, almost every boot, I create a /tmp/t, and then I pollute
it.
With this, I can make auto-cleaned-up directories that are arbitrary
(`tt`) that's for the quick experiments.
And then there's a helper for longer term stuff (via `t`) that's
equivalent to what I already do, but nicer.